requirements.txt
cython
scipy
shapely
timm
h5py
wandb
fire
opencv-python
openai-clip
pandas
braceexpand
torch-ema
torchmetrics
setuptools
webdataset>=0.2.5